Started Timer ESP32 redesign
This redesign replaces the Timer's ATTINY88 with an ESP32-C3 to make it more compatible with the other flight computers. It also adds a USB-C port for easier programming.

The Timer revision was started by copying the Guesser and removing everything unnecessary, then yoinking the ESP32-C3 off of the Pointer schematic, so everything is currently kinda a huge mess and some stuff just isn't connected together correctly at all, but all of that will be fixed soon.

Also it's currently in the "the_timer_2" folder so the old timer stuff is still referenceable. This will be changed later once I have unexploded the schematic and started the PCB design.